requirements.txt
Flask==3.0.0
supabase==2.27.0
python-dotenv==1.0.0
python-dateutil==2.8.2
gunicorn==21.2.0